added the logic for run

+++ AiPlayground.DataAccess/Entities/Run.cs
@@ -4,9 +4,9 @@ public class Run
 {
     public int Id { get; set; }
     public string ActualResponse { get; set; } = string.Empty;
-    public decimal Rating { get; set; }
-    public decimal UserRating { get; set; }
-    public decimal Temp { get; set; }
+    public double Rating { get; set; }
+    public double UserRating { get; set; }
+    public double Temp { get; set; }
     
     public int PromptId { get; set; }
     public int ModelId { get; set; }
